Move cid generator to a module and export trait
