+++ quinn-proto/src/lib.rs
@@ -59,7 +59,10 @@ pub use crate::transport_error::{Code as TransportErrorCode, Error as TransportE
 
 pub mod congestion;
 
+mod cid_generator;
 mod token;
+pub use crate::cid_generator::ConnectionIdGenerator;
+
 use token::{ResetToken, RetryToken};
 
 /// Types that are generic over the crypto protocol implementation
